Extract OTLP log test helpers
Move log-specific OTLP test builders into OtlpLogUtils so OtlpUtils
stays focused on shared and metrics helpers.
